@@ -1,8 +1,19 @@
 package no.vestlandetmc.BanFromClaim;
 
-import java.io.File;
-import java.io.IOException;
-
+import no.vestlandetmc.BanFromClaim.commands.SafeSpot;
+import no.vestlandetmc.BanFromClaim.commands.griefdefender.*;
+import no.vestlandetmc.BanFromClaim.commands.griefprevention.*;
+import no.vestlandetmc.BanFromClaim.config.ClaimData;
+import no.vestlandetmc.BanFromClaim.config.Config;
+import no.vestlandetmc.BanFromClaim.config.Messages;
+import no.vestlandetmc.BanFromClaim.handler.Hooks;
+import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
+import no.vestlandetmc.BanFromClaim.listener.CombatMode;
+import no.vestlandetmc.BanFromClaim.listener.GDListener;
+import no.vestlandetmc.BanFromClaim.listener.GPListener;
+import no.vestlandetmc.BanFromClaim.listener.PlayerListener;
+import no.vestlandetmc.BanFromClaim.schedulers.CombatScheduler;
+import no.vestlandetmc.BanFromClaim.utils.UpdateNotification;
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.InvalidConfigurationException;
 import org.bukkit.configuration.file.FileConfiguration;
@@ -10,33 +21,8 @@ import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.scheduler.BukkitRunnable;
 
-import no.vestlandetmc.BanFromClaim.commands.SafeSpot;
-import no.vestlandetmc.BanFromClaim.commands.griefdefender.BfcAllCommandGD;
-import no.vestlandetmc.BanFromClaim.commands.griefdefender.BfcCommandGD;
-import no.vestlandetmc.BanFromClaim.commands.griefdefender.BfclistCommandGD;
-import no.vestlandetmc.BanFromClaim.commands.griefdefender.KfcCommandGD;
-import no.vestlandetmc.BanFromClaim.commands.griefdefender.UnbfcCommandGD;
-import no.vestlandetmc.BanFromClaim.commands.griefprevention.BfcAllCommand;
-import no.vestlandetmc.BanFromClaim.commands.griefprevention.BfcCommand;
-import no.vestlandetmc.BanFromClaim.commands.griefprevention.BfclistCommand;
-import no.vestlandetmc.BanFromClaim.commands.griefprevention.KfcCommandGP;
-import no.vestlandetmc.BanFromClaim.commands.griefprevention.UnbfcCommand;
-import no.vestlandetmc.BanFromClaim.commands.regiondefence.BfcAllCommandRD;
-import no.vestlandetmc.BanFromClaim.commands.regiondefence.BfcCommandRD;
-import no.vestlandetmc.BanFromClaim.commands.regiondefence.BfclistCommandRD;
-import no.vestlandetmc.BanFromClaim.commands.regiondefence.UnbfcCommandRD;
-import no.vestlandetmc.BanFromClaim.config.ClaimData;
-import no.vestlandetmc.BanFromClaim.config.Config;
-import no.vestlandetmc.BanFromClaim.config.Messages;
-import no.vestlandetmc.BanFromClaim.handler.CombatScheduler;
-import no.vestlandetmc.BanFromClaim.handler.Hooks;
-import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
-import no.vestlandetmc.BanFromClaim.handler.UpdateNotification;
-import no.vestlandetmc.BanFromClaim.listener.CombatMode;
-import no.vestlandetmc.BanFromClaim.listener.GDListener;
-import no.vestlandetmc.BanFromClaim.listener.GPListener;
-import no.vestlandetmc.BanFromClaim.listener.PlayerListener;
-import no.vestlandetmc.BanFromClaim.listener.RDListener;
+import java.io.File;
+import java.io.IOException;
 
 public class BfcPlugin extends JavaPlugin {
 
@@ -61,7 +47,7 @@ public class BfcPlugin extends JavaPlugin {
 
 		Config.initialize();
 
-		if(getServer().getPluginManager().getPlugin("GriefPrevention") != null) {
+		if (getServer().getPluginManager().getPlugin("GriefPrevention") != null) {
 			MessageHandler.sendConsole("&2[" + getDescription().getPrefix() + "] &7Successfully hooked into &eGriefPrevention");
 			MessageHandler.sendConsole("");
 
@@ -73,11 +59,11 @@ public class BfcPlugin extends JavaPlugin {
 			this.getCommand("banfromclaimlist").setExecutor(new BfclistCommand());
 			this.getCommand("banfromclaimall").setExecutor(new BfcAllCommand());
 
-			if(Config.KICKMODE) {
+			if (Config.KICKMODE) {
 				this.getCommand("kickfromclaim").setExecutor(new KfcCommandGP());
 			}
 
-		} else if(getServer().getPluginManager().getPlugin("GriefDefender") != null) {
+		} else if (getServer().getPluginManager().getPlugin("GriefDefender") != null) {
 			MessageHandler.sendConsole("&2[" + getDescription().getPrefix() + "] &7Successfully hooked into &eGriefDefender");
 			MessageHandler.sendConsole("");
 
@@ -89,23 +75,7 @@ public class BfcPlugin extends JavaPlugin {
 			this.getCommand("banfromclaimlist").setExecutor(new BfclistCommandGD());
 			this.getCommand("banfromclaimall").setExecutor(new BfcAllCommandGD());
 
-			if(Config.KICKMODE) {
-				this.getCommand("kickfromclaim").setExecutor(new KfcCommandGD());
-			}
-
-		} else if(getServer().getPluginManager().getPlugin("RegionDefence") != null) {
-			MessageHandler.sendConsole("&2[" + getDescription().getPrefix() + "] &7Successfully hooked into &eRegionDefence");
-			MessageHandler.sendConsole("");
-
-			Hooks.setRD();
-
-			this.getServer().getPluginManager().registerEvents(new RDListener(), this);
-			this.getCommand("banfromclaim").setExecutor(new BfcCommandRD());
-			this.getCommand("unbanfromclaim").setExecutor(new UnbfcCommandRD());
-			this.getCommand("banfromclaimlist").setExecutor(new BfclistCommandRD());
-			this.getCommand("banfromclaimall").setExecutor(new BfcAllCommandRD());
-
-			if(Config.KICKMODE) {
+			if (Config.KICKMODE) {
 				this.getCommand("kickfromclaim").setExecutor(new KfcCommandGD());
 			}
 
@@ -123,7 +93,7 @@ public class BfcPlugin extends JavaPlugin {
 		Messages.initialize();
 		ClaimData.createSection();
 
-		if(Config.COMBAT_ENABLED) {
+		if (Config.COMBAT_ENABLED) {
 			this.getServer().getPluginManager().registerEvents(new CombatMode(), this);
 			new CombatScheduler().runTaskTimer(this, 0L, 20L);
 		}
@@ -148,11 +118,6 @@ public class BfcPlugin extends JavaPlugin {
 		}.runTaskAsynchronously(this);
 	}
 
-	@Override
-	public void onDisable() {
-
-	}
-
 	public FileConfiguration getDataFile() {
 		return this.data;
 	}
@@ -164,7 +129,7 @@ public class BfcPlugin extends JavaPlugin {
 			try {
 				dataFile.createNewFile();
 			} catch (final IOException e) {
-				e.printStackTrace();
+				getLogger().severe(e.getMessage());
 			}
 		}
 
@@ -172,7 +137,7 @@ public class BfcPlugin extends JavaPlugin {
 		try {
 			data.load(dataFile);
 		} catch (IOException | InvalidConfigurationException e) {
-			e.printStackTrace();
+			getLogger().severe(e.getMessage());
 		}
 	}
 }
