@@ -1,26 +1,24 @@
 package no.vestlandetmc.BanFromClaim.commands.griefdefender;
 
+import com.griefdefender.api.Core;
+import com.griefdefender.api.GriefDefender;
+import com.griefdefender.api.claim.Claim;
+import com.griefdefender.api.claim.TrustTypes;
+import com.griefdefender.lib.flowpowered.math.vector.Vector3i;
+import no.vestlandetmc.BanFromClaim.config.ClaimData;
+import no.vestlandetmc.BanFromClaim.config.Messages;
+import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
 import org.bukkit.Location;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-import com.griefdefender.api.Core;
-import com.griefdefender.api.GriefDefender;
-import com.griefdefender.api.claim.Claim;
-import com.griefdefender.api.claim.TrustTypes;
-import com.griefdefender.lib.flowpowered.math.vector.Vector3i;
-
-import no.vestlandetmc.BanFromClaim.config.ClaimData;
-import no.vestlandetmc.BanFromClaim.config.Messages;
-import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
-
 public class BfcAllCommandGD implements CommandExecutor {
 
 	@Override
 	public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-		if(!(sender instanceof Player)) {
+		if (!(sender instanceof Player)) {
 			MessageHandler.sendConsole("&cThis command can only be used in-game.");
 			return true;
 		}
@@ -36,20 +34,25 @@ public class BfcAllCommandGD implements CommandExecutor {
 		final boolean isOwner = claim.getOwnerUniqueId().equals(player.getUniqueId());
 		boolean allowBan = false;
 
-		if(claim.isWilderness()) {
+		if (claim.isWilderness()) {
 			MessageHandler.sendMessage(player, Messages.OUTSIDE_CLAIM);
 			return true;
 		}
 
-		if(isOwner || isManager) { allowBan = true; }
-		else if(player.hasPermission("bfc.admin")) { allowBan = true; }
+		if (isOwner || isManager) {
+			allowBan = true;
+		} else if (player.hasPermission("bfc.admin")) {
+			allowBan = true;
+		}
 
-		if(allowBan) {
+		if (allowBan) {
 			claimData.banAll(claim.getUniqueId().toString());
 
-			if(claimData.isAllBanned(claim.getUniqueId().toString())) {
+			if (claimData.isAllBanned(claim.getUniqueId().toString())) {
 				MessageHandler.sendMessage(player, Messages.BAN_ALL);
-			} else { MessageHandler.sendMessage(player, Messages.UNBAN_ALL); }
+			} else {
+				MessageHandler.sendMessage(player, Messages.UNBAN_ALL);
+			}
 
 		} else {
 			MessageHandler.sendMessage(player, Messages.NO_ACCESS);
