@@ -1,8 +1,13 @@
 package no.vestlandetmc.BanFromClaim.commands.griefdefender;
 
-import java.util.List;
-import java.util.UUID;
-
+import com.griefdefender.api.Core;
+import com.griefdefender.api.GriefDefender;
+import com.griefdefender.api.claim.Claim;
+import com.griefdefender.api.claim.TrustTypes;
+import com.griefdefender.lib.flowpowered.math.vector.Vector3i;
+import no.vestlandetmc.BanFromClaim.config.ClaimData;
+import no.vestlandetmc.BanFromClaim.config.Messages;
+import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.OfflinePlayer;
@@ -11,15 +16,8 @@ import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-import com.griefdefender.api.Core;
-import com.griefdefender.api.GriefDefender;
-import com.griefdefender.api.claim.Claim;
-import com.griefdefender.api.claim.TrustTypes;
-import com.griefdefender.lib.flowpowered.math.vector.Vector3i;
-
-import no.vestlandetmc.BanFromClaim.config.ClaimData;
-import no.vestlandetmc.BanFromClaim.config.Messages;
-import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
+import java.util.List;
+import java.util.UUID;
 
 public class BfclistCommandGD implements CommandExecutor {
 
@@ -29,7 +27,7 @@ public class BfclistCommandGD implements CommandExecutor {
 
 	@Override
 	public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-		if(!(sender instanceof Player)) {
+		if (!(sender instanceof Player)) {
 			MessageHandler.sendConsole("&cThis command can only be used in-game.");
 			return true;
 		}
@@ -41,19 +39,18 @@ public class BfclistCommandGD implements CommandExecutor {
 		final Claim claim = gd.getClaimManager(loc.getWorld().getUID()).getClaimAt(vector);
 		final ClaimData claimData = new ClaimData();
 
-		if(args.length != 0) {
-			if(isInt(args[0])) {
+		if (args.length != 0) {
+			if (isInt(args[0])) {
 				this.number = Integer.parseInt(args[0]);
 				this.countTo = 5 * number;
 				this.countFrom = 5 * number - 5;
-			}
-			else {
+			} else {
 				MessageHandler.sendMessage(player, Messages.UNVALID_NUMBER);
 				return true;
 			}
 		}
 
-		if(claim.isWilderness()) {
+		if (claim.isWilderness()) {
 			MessageHandler.sendMessage(player, Messages.OUTSIDE_CLAIM);
 			return true;
 		}
@@ -62,51 +59,53 @@ public class BfclistCommandGD implements CommandExecutor {
 		final boolean isOwner = claim.getOwnerUniqueId().equals(player.getUniqueId());
 		boolean allowBan = false;
 
-		if(isOwner || isManager) { allowBan = true; }
-		else if(player.hasPermission("bfc.admin")) { allowBan = true; }
+		if (isOwner || isManager) {
+			allowBan = true;
+		} else if (player.hasPermission("bfc.admin")) {
+			allowBan = true;
+		}
 
-		int totalPage = 1;
+		int totalPage;
 
-		if(!allowBan) {
+		if (!allowBan) {
 			MessageHandler.sendMessage(player, Messages.NO_ACCESS);
 			return true;
 
 		} else {
 			MessageHandler.sendMessage(player, Messages.placeholders(Messages.LIST_HEADER, null, player.getDisplayName(), claim.getOwnerName()));
 
-			if(claimData.isAllBanned(claim.getUniqueId().toString())) {
+			if (claimData.isAllBanned(claim.getUniqueId().toString())) {
 				MessageHandler.sendMessage(player, Messages.LIST_BAN_ALL);
 				return true;
 			}
 
-			if(listPlayers(claim.getUniqueId().toString()) == null) {
+			if (listPlayers(claim.getUniqueId().toString()) == null) {
 				MessageHandler.sendMessage(player, Messages.placeholders(Messages.LIST_EMPTY, null, player.getDisplayName(), claim.getOwnerName()));
 				return true;
 			} else {
 				totalPage = listPlayers(claim.getUniqueId().toString()).size() / 5 + 1;
-				for(int i = 0; i < listPlayers(claim.getUniqueId().toString()).toArray().length; i++) {
-					if(this.number > totalPage || this.number == 0) {
+				for (int i = 0; i < listPlayers(claim.getUniqueId().toString()).toArray().length; i++) {
+					if (this.number > totalPage || this.number == 0) {
 						this.countTo = 5 * totalPage;
 						this.countFrom = 5 * totalPage - 5;
 						this.number = totalPage;
 					}
 
-					if(i >= this.countFrom) {
+					if (i >= this.countFrom) {
 						final String bp = (String) listPlayers(claim.getUniqueId().toString()).toArray()[i];
 						final OfflinePlayer bannedPlayer = Bukkit.getOfflinePlayer(UUID.fromString(bp));
 						MessageHandler.sendMessage(player, "&6" + bannedPlayer.getName());
 
-						if(i == this.countTo) {
+						if (i == this.countTo) {
 							MessageHandler.sendMessage(player, "");
 							MessageHandler.sendMessage(player, "&e<--- [&6" + this.number + "\\" + totalPage + "&e] --->");
 							break;
 						}
 
-						continue;
 					}
 				}
 
-				if(this.number == totalPage) {
+				if (this.number == totalPage) {
 					MessageHandler.sendMessage(player, "");
 					MessageHandler.sendMessage(player, "&e<--- [&6" + totalPage + "\\" + totalPage + "&e] --->");
 				}
