@@ -1,8 +1,10 @@
 package no.vestlandetmc.BanFromClaim.commands.griefprevention;
 
-import java.util.List;
-import java.util.UUID;
-
+import me.ryanhamshire.GriefPrevention.Claim;
+import me.ryanhamshire.GriefPrevention.GriefPrevention;
+import no.vestlandetmc.BanFromClaim.config.ClaimData;
+import no.vestlandetmc.BanFromClaim.config.Messages;
+import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.OfflinePlayer;
@@ -11,18 +13,15 @@ import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-import me.ryanhamshire.GriefPrevention.Claim;
-import me.ryanhamshire.GriefPrevention.GriefPrevention;
-import no.vestlandetmc.BanFromClaim.config.ClaimData;
-import no.vestlandetmc.BanFromClaim.config.Messages;
-import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
+import java.util.List;
+import java.util.UUID;
 
 public class UnbfcCommand implements CommandExecutor {
 
 	@SuppressWarnings("deprecation")
 	@Override
 	public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-		if(!(sender instanceof Player)) {
+		if (!(sender instanceof Player)) {
 			MessageHandler.sendConsole("&cThis command can only be used in-game.");
 			return true;
 		}
@@ -31,25 +30,26 @@ public class UnbfcCommand implements CommandExecutor {
 		final Location loc = player.getLocation();
 		final Claim claim = GriefPrevention.instance.dataStore.getClaimAt(loc, true, null);
 
-		if(args.length == 0) {
+		if (args.length == 0) {
 			MessageHandler.sendMessage(player, Messages.NO_ARGUMENTS);
 			return true;
 		}
 
-		if(claim == null) {
+		if (claim == null) {
 			MessageHandler.sendMessage(player, Messages.OUTSIDE_CLAIM);
 			return true;
 		}
 
 		final String accessDenied = claim.allowGrantPermission(player);
-		boolean allowBan = false;
+		boolean allowBan = accessDenied == null;
 
-		if(accessDenied == null) { allowBan = true; }
-		if(player.hasPermission("bfc.admin")) { allowBan = true; }
+		if (player.hasPermission("bfc.admin")) {
+			allowBan = true;
+		}
 
 		OfflinePlayer bPlayer = null;
 
-		if(!allowBan) {
+		if (!allowBan) {
 			MessageHandler.sendMessage(player, Messages.NO_ACCESS);
 			return true;
 
@@ -57,14 +57,14 @@ public class UnbfcCommand implements CommandExecutor {
 			final String claimOwner = claim.getOwnerName();
 			final String claimID = claim.getID().toString();
 
-			if(listPlayers(claim.getID().toString()) != null) {
-				for(final String bp : listPlayers(claim.getID().toString())) {
+			if (listPlayers(claim.getID().toString()) != null) {
+				for (final String bp : listPlayers(claim.getID().toString())) {
 					final OfflinePlayer bannedPlayer = Bukkit.getOfflinePlayer(UUID.fromString(bp));
-					if(bannedPlayer.getName().equalsIgnoreCase(args[0])) {
+					if (bannedPlayer.getName().equalsIgnoreCase(args[0])) {
 						bPlayer = bannedPlayer;
-						if(setClaimData(player, claimID, bp, false)) {
+						if (setClaimData(player, claimID, bp, false)) {
 							MessageHandler.sendMessage(player, Messages.placeholders(Messages.UNBANNED, bannedPlayer.getName(), player.getDisplayName(), claimOwner));
-							if(bannedPlayer.isOnline()) {
+							if (bannedPlayer.isOnline()) {
 								MessageHandler.sendMessage(bannedPlayer.getPlayer(), Messages.placeholders(Messages.UNBANNED_TARGET, bannedPlayer.getName(), player.getDisplayName(), claimOwner));
 							}
 							return true;
@@ -74,7 +74,9 @@ public class UnbfcCommand implements CommandExecutor {
 			}
 		}
 
-		if(bPlayer == null) { MessageHandler.sendMessage(player, Messages.placeholders(Messages.NOT_BANNED, args[0], player.getDisplayName(), null)); }
+		if (bPlayer == null) {
+			MessageHandler.sendMessage(player, Messages.placeholders(Messages.NOT_BANNED, args[0], player.getDisplayName(), null));
+		}
 
 		return true;
 	}
