@@ -1,21 +1,20 @@
 package no.vestlandetmc.BanFromClaim.config;
 
+import com.griefdefender.api.Core;
+import com.griefdefender.api.GriefDefender;
+import com.griefdefender.api.claim.Claim;
+import me.ryanhamshire.GriefPrevention.GriefPrevention;
+import no.vestlandetmc.BanFromClaim.BfcPlugin;
+import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
+import org.bukkit.Bukkit;
+import org.bukkit.configuration.file.FileConfiguration;
+import org.bukkit.entity.Player;
+
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
-
-import org.bukkit.configuration.file.FileConfiguration;
-import org.bukkit.entity.Player;
-
-import com.griefdefender.api.Core;
-import com.griefdefender.api.GriefDefender;
-import com.griefdefender.api.claim.Claim;
-
-import me.ryanhamshire.GriefPrevention.GriefPrevention;
-import no.vestlandetmc.BanFromClaim.BfcPlugin;
-import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
 
 public class ClaimData {
 
@@ -28,81 +27,86 @@ public class ClaimData {
 	}
 
 	public boolean setClaimData(Player player, String claimID, String bannedUUID, boolean add) {
-		if(add) {
-			if(!existData(claimID, bannedUUID)) {
+		if (add) {
+			if (!existData(claimID, bannedUUID)) {
 				addData(claimID, bannedUUID);
 				return true;
-			} else { return false; }
+			} else {
+				return false;
+			}
 		} else {
-			if(existData(claimID, bannedUUID)) {
+			if (existData(claimID, bannedUUID)) {
 				removeData(claimID, bannedUUID);
 				return true;
-			} else { return false; }
+			} else {
+				return false;
+			}
 		}
 	}
 
 	private void addData(String claimID, String bannedUUID) {
 		final List<String> uuid = new ArrayList<>();
 
-		if(!cfg.contains(prefix + "." + claimID)) { cfg.createSection(prefix + "." + claimID); }
-
-		if(cfg.getStringList(prefix + "." + claimID).isEmpty()) {
-			uuid.add(bannedUUID);
-			cfg.set(prefix + "." + claimID, uuid);
-			saveDatafile();
-		} else {
-			uuid.addAll(cfg.getStringList(prefix + "." + claimID));
-			uuid.add(bannedUUID);
-			cfg.set(prefix + "." + claimID, uuid);
-			saveDatafile();
+		if (!cfg.contains(prefix + "." + claimID)) {
+			cfg.createSection(prefix + "." + claimID);
 		}
+
+		if (!cfg.getStringList(prefix + "." + claimID).isEmpty()) {
+			uuid.addAll(cfg.getStringList(prefix + "." + claimID));
+		}
+
+		uuid.add(bannedUUID);
+		cfg.set(prefix + "." + claimID, uuid);
+		saveDatafile();
 	}
 
 	public void banAll(String claimID) {
-		if(cfg.contains("claims-ban-all" + "." + claimID + ".ban-all")) {
-			if(cfg.getBoolean("claims-ban-all" + "." + claimID + ".ban-all")) {
+		if (cfg.contains("claims-ban-all" + "." + claimID + ".ban-all")) {
+			if (cfg.getBoolean("claims-ban-all" + "." + claimID + ".ban-all")) {
 				cfg.set("claims-ban-all" + "." + claimID + ".ban-all", false);
-			} else { cfg.set("claims-ban-all" + "." + claimID + ".ban-all", true); }
-		} else { cfg.set("claims-ban-all" + "." + claimID + ".ban-all", true); }
+			} else {
+				cfg.set("claims-ban-all" + "." + claimID + ".ban-all", true);
+			}
+		} else {
+			cfg.set("claims-ban-all" + "." + claimID + ".ban-all", true);
+		}
 
 		saveDatafile();
 	}
 
 	public boolean isAllBanned(String claimID) {
-		if(cfg.contains("claims-ban-all" + "." + claimID + ".ban-all")) {
+		if (cfg.contains("claims-ban-all" + "." + claimID + ".ban-all")) {
 			return cfg.getBoolean("claims-ban-all" + "." + claimID + ".ban-all");
-		} else { return false; }
+		} else {
+			return false;
+		}
 	}
 
 	private void removeData(String claimID, String bannedUUID) {
-		final List<String> uuid = new ArrayList<>();
 
-		if(!cfg.getStringList(prefix + "." + claimID).isEmpty()) {
-			uuid.addAll(cfg.getStringList(prefix + "." + claimID));
-			if(uuid.contains(bannedUUID)) {
+		if (!cfg.getStringList(prefix + "." + claimID).isEmpty()) {
+			final List<String> uuid = new ArrayList<>(cfg.getStringList(prefix + "." + claimID));
+			if (uuid.contains(bannedUUID)) {
 				uuid.remove(bannedUUID);
 				cfg.set(prefix + "." + claimID, uuid);
 
-				if(cfg.getStringList(prefix + "." + claimID).isEmpty()) { cfg.set(prefix + "." + claimID, null); }
+				if (cfg.getStringList(prefix + "." + claimID).isEmpty()) {
+					cfg.set(prefix + "." + claimID, null);
+				}
 				saveDatafile();
 			}
 		}
 	}
 
 	private boolean existData(String claimID, String bannedUUID) {
-		final List<String> uuid = new ArrayList<>();
 
-		if(cfg.contains(prefix + "." + claimID)) {
-			if(cfg.getStringList(prefix + "." + claimID).isEmpty()) {
+		if (cfg.contains(prefix + "." + claimID)) {
+			if (cfg.getStringList(prefix + "." + claimID).isEmpty()) {
 				return false;
 
 			} else {
-				uuid.addAll(cfg.getStringList(prefix + "." + claimID));
-				if(uuid.contains(bannedUUID)) {
-					return true;
-				} else {
-					return false;
-				}
+				final List<String> uuid = new ArrayList<>(cfg.getStringList(prefix + "." + claimID));
+				return uuid.contains(bannedUUID);
 			}
 		}
 
@@ -110,16 +114,12 @@ public class ClaimData {
 	}
 
 	public boolean checkClaim(String claimID) {
-		if(cfg.contains(prefix + "." + claimID)) {
-			return true;
-		} else {
-			return false;
-		}
+		return cfg.contains(prefix + "." + claimID);
 	}
 
 	public List<String> bannedPlayers(String claimID) {
-		if(cfg.contains(prefix + "." + claimID)) {
-			if(!cfg.getStringList(prefix + "." + claimID).isEmpty()) {
+		if (cfg.contains(prefix + "." + claimID)) {
+			if (!cfg.getStringList(prefix + "." + claimID).isEmpty()) {
 				return cfg.getStringList(prefix + "." + claimID);
 			}
 		}
@@ -132,13 +132,17 @@ public class ClaimData {
 			file = new File(BfcPlugin.getInstance().getDataFolder(), "data.dat");
 			BfcPlugin.getInstance().getDataFile().save(file);
 		} catch (final IOException e) {
-			e.printStackTrace();
+			Bukkit.getLogger().severe(e.getMessage());
 		}
 	}
 
 	public static void createSection() {
-		if(!BfcPlugin.getInstance().getDataFile().contains("bfc_claim_data")) { BfcPlugin.getInstance().getDataFile().createSection("bfc_claim_data"); }
-		if(!BfcPlugin.getInstance().getDataFile().contains("claims-ban-all")) { BfcPlugin.getInstance().getDataFile().createSection("claims-ban-all"); }
+		if (!BfcPlugin.getInstance().getDataFile().contains("bfc_claim_data")) {
+			BfcPlugin.getInstance().getDataFile().createSection("bfc_claim_data");
+		}
+		if (!BfcPlugin.getInstance().getDataFile().contains("claims-ban-all")) {
+			BfcPlugin.getInstance().getDataFile().createSection("claims-ban-all");
+		}
 		saveDatafile();
 	}
 
@@ -146,22 +150,20 @@ public class ClaimData {
 		boolean clean = false;
 		final String prefix = BfcPlugin.getInstance().getDescription().getPrefix();
 
-		if(!BfcPlugin.getInstance().getDataFile().getKeys(false).isEmpty()) {
-			if(!BfcPlugin.getInstance().getDataFile().getConfigurationSection("bfc_claim_data").getKeys(false).isEmpty()) {
-				for(final String claimID : BfcPlugin.getInstance().getDataFile().getConfigurationSection("bfc_claim_data").getKeys(false)) {
-					if(BfcPlugin.getInstance().getServer().getPluginManager().getPlugin("GriefPrevention") != null) {
-						if(GriefPrevention.instance.dataStore.getClaim(Long.parseLong(claimID)) == null) {
+		if (!BfcPlugin.getInstance().getDataFile().getKeys(false).isEmpty()) {
+			if (!BfcPlugin.getInstance().getDataFile().getConfigurationSection("bfc_claim_data").getKeys(false).isEmpty()) {
+				for (final String claimID : BfcPlugin.getInstance().getDataFile().getConfigurationSection("bfc_claim_data").getKeys(false)) {
+					if (BfcPlugin.getInstance().getServer().getPluginManager().getPlugin("GriefPrevention") != null) {
+						if (GriefPrevention.instance.dataStore.getClaim(Long.parseLong(claimID)) == null) {
 							BfcPlugin.getInstance().getDataFile().set("bfc_claim_data." + claimID, null);
 							clean = true;
 						}
-					}
-
-					else if(BfcPlugin.getInstance().getServer().getPluginManager().getPlugin("GriefDefender") != null) {
+					} else if (BfcPlugin.getInstance().getServer().getPluginManager().getPlugin("GriefDefender") != null) {
 						final Core gd = GriefDefender.getCore();
 						final UUID uuid = UUID.fromString(claimID);
 						final Claim claim = gd.getClaim(uuid);
 
-						if(claim == null) {
+						if (claim == null) {
 							BfcPlugin.getInstance().getDataFile().set("bfc_claim_data." + claimID, null);
 							clean = true;
 						}
@@ -169,21 +171,19 @@ public class ClaimData {
 				}
 			}
 
-			if(!BfcPlugin.getInstance().getDataFile().getConfigurationSection("claims-ban-all").getKeys(false).isEmpty()) {
-				for(final String claimID : BfcPlugin.getInstance().getDataFile().getConfigurationSection("bfc_claim_data").getKeys(false)) {
-					if(BfcPlugin.getInstance().getServer().getPluginManager().getPlugin("GriefPrevention") != null) {
-						if(GriefPrevention.instance.dataStore.getClaim(Long.parseLong(claimID)) == null) {
+			if (!BfcPlugin.getInstance().getDataFile().getConfigurationSection("claims-ban-all").getKeys(false).isEmpty()) {
+				for (final String claimID : BfcPlugin.getInstance().getDataFile().getConfigurationSection("bfc_claim_data").getKeys(false)) {
+					if (BfcPlugin.getInstance().getServer().getPluginManager().getPlugin("GriefPrevention") != null) {
+						if (GriefPrevention.instance.dataStore.getClaim(Long.parseLong(claimID)) == null) {
 							BfcPlugin.getInstance().getDataFile().set("claims-ban-all." + claimID, null);
 							clean = true;
 						}
-					}
-
-					else if(BfcPlugin.getInstance().getServer().getPluginManager().getPlugin("GriefDefender") != null) {
+					} else if (BfcPlugin.getInstance().getServer().getPluginManager().getPlugin("GriefDefender") != null) {
 						final Core gd = GriefDefender.getCore();
 						final UUID uuid = UUID.fromString(claimID);
 						final Claim claim = gd.getClaim(uuid);
 
-						if(claim == null) {
+						if (claim == null) {
 							BfcPlugin.getInstance().getDataFile().set("claims-ban-all." + claimID, null);
 							clean = true;
 						}
@@ -192,7 +192,7 @@ public class ClaimData {
 			}
 		}
 
-		if(clean) {
+		if (clean) {
 			saveDatafile();
 			MessageHandler.sendConsole("&2[" + prefix + "] &eData storage has been cleared of old removed claims...");
 		}
