@@ -4,7 +4,7 @@ public class Hooks {
 
 	private static boolean griefPrevention = false;
 	private static boolean griefDefender = false;
-	private static boolean regionDefence = false;
+	private static final boolean gsit = false;
 
 	public static boolean gpEnabled() {
 		return griefPrevention;
@@ -14,20 +14,12 @@ public class Hooks {
 		return griefDefender;
 	}
 
-	public static boolean rdEnabled() {
-		return regionDefence;
-	}
-
 	public static void setGP() {
 		griefPrevention = true;
 	}
 
 	public static void setGD() {
 		griefDefender = true;
-	}
-
-	public static void setRD() {
-		regionDefence = true;
 	}
 
 }
