@@ -1,14 +1,13 @@
 package no.vestlandetmc.BanFromClaim.handler;
 
-import java.util.ArrayList;
-
+import net.md_5.bungee.api.ChatMessageType;
+import net.md_5.bungee.api.chat.TextComponent;
+import no.vestlandetmc.BanFromClaim.BfcPlugin;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 
-import net.md_5.bungee.api.ChatMessageType;
-import net.md_5.bungee.api.chat.TextComponent;
-import no.vestlandetmc.BanFromClaim.BfcPlugin;
+import java.util.ArrayList;
 
 public class MessageHandler {
 
@@ -27,21 +26,21 @@ public class MessageHandler {
 	}
 
 	public static void sendMessage(Player player, String... messages) {
-		for(final String message : messages) {
+		for (final String message : messages) {
 			player.sendMessage(colorize(message));
 		}
 	}
 
 	public static void sendAnnounce(String... messages) {
-		for(final Player player : Bukkit.getOnlinePlayers()) {
-			for(final String message : messages) {
+		for (final Player player : Bukkit.getOnlinePlayers()) {
+			for (final String message : messages) {
 				player.sendMessage(colorize(message));
 			}
 		}
 	}
 
 	public static void sendConsole(String... messages) {
-		for(final String message : messages) {
+		for (final String message : messages) {
 			BfcPlugin.getInstance().getServer().getConsoleSender().sendMessage(colorize(message));
 		}
 	}
@@ -51,13 +50,12 @@ public class MessageHandler {
 	}
 
 	public static String placeholders(String message, String time, String tpName, String tp1, String tp2) {
-		final String converted = message.
+
+		return message.
 				replaceAll("%time%", time).
 				replaceAll("%tpname%", tpName).
 				replaceAll("%teleport1%", tp1).
 				replaceAll("%teleport2%", tp2);
-
-		return converted;
 
 	}
 
