@@ -1,7 +1,18 @@
 package no.vestlandetmc.BanFromClaim.listener;
 
-import java.util.UUID;
-
+import com.griefdefender.api.Core;
+import com.griefdefender.api.GriefDefender;
+import com.griefdefender.api.claim.Claim;
+import com.griefdefender.api.claim.TrustTypes;
+import com.griefdefender.lib.flowpowered.math.vector.Vector3i;
+import no.vestlandetmc.BanFromClaim.BfcPlugin;
+import no.vestlandetmc.BanFromClaim.config.ClaimData;
+import no.vestlandetmc.BanFromClaim.config.Config;
+import no.vestlandetmc.BanFromClaim.config.Messages;
+import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
+import no.vestlandetmc.BanFromClaim.handler.ParticleHandler;
+import no.vestlandetmc.BanFromClaim.utils.LocationFinder;
+import no.vestlandetmc.BanFromClaim.utils.PlayerRidePlayer;
 import org.bukkit.Bukkit;
 import org.bukkit.GameMode;
 import org.bukkit.Location;
@@ -11,19 +22,7 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerMoveEvent;
 
-import com.griefdefender.api.Core;
-import com.griefdefender.api.GriefDefender;
-import com.griefdefender.api.claim.Claim;
-import com.griefdefender.api.claim.TrustTypes;
-import com.griefdefender.lib.flowpowered.math.vector.Vector3i;
-
-import no.vestlandetmc.BanFromClaim.BfcPlugin;
-import no.vestlandetmc.BanFromClaim.config.ClaimData;
-import no.vestlandetmc.BanFromClaim.config.Config;
-import no.vestlandetmc.BanFromClaim.config.Messages;
-import no.vestlandetmc.BanFromClaim.handler.LocationFinder;
-import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
-import no.vestlandetmc.BanFromClaim.handler.ParticleHandler;
+import java.util.UUID;
 
 public class GDListener implements Listener {
 
@@ -33,7 +32,9 @@ public class GDListener implements Listener {
 		final Location locFrom = e.getFrom();
 		final Location locTo = e.getTo();
 
-		if(locFrom.getBlock().equals(locTo.getBlock())) { return; }
+		if (locFrom.getBlock().equals(locTo.getBlock())) {
+			return;
+		}
 
 		final Player player = e.getPlayer();
 		final Core gd = GriefDefender.getCore();
@@ -43,19 +44,26 @@ public class GDListener implements Listener {
 		final Claim claimFrom = gd.getClaimManager(locFrom.getWorld().getUID()).getClaimAt(vectorFrom);
 		final ParticleHandler ph = new ParticleHandler(e.getTo());
 
-		if(locFrom.getBlockX() != locTo.getBlockX() || locFrom.getBlockZ() != locTo.getBlockZ()) {
-			if(player.hasPermission("bfc.bypass") || player.getGameMode().equals(GameMode.SPECTATOR)) { return; }
-
-			if(!claimTo.isWilderness()) {
+		if (locFrom.getBlockX() != locTo.getBlockX() || locFrom.getBlockZ() != locTo.getBlockZ()) {
+			if (!claimTo.isWilderness()) {
 				final UUID ownerUUID = claimTo.getOwnerUniqueId();
+				final Player target = PlayerRidePlayer.getPassenger(player);
 				boolean hasAttacked = false;
 
-				if(CombatMode.attackerContains(player.getUniqueId()))
+				if (target != null && (claimData.isAllBanned(claimTo.getUniqueId().toString()) || playerBanned(target, claimTo) || playerBanned(player, claimTo))) {
+					target.teleport(player.getLocation().add(0, 4, 0));
+				}
+
+				if (CombatMode.attackerContains(player.getUniqueId()))
 					hasAttacked = CombatMode.getAttacker(player.getUniqueId()).equals(ownerUUID);
 
-				if((claimData.isAllBanned(claimTo.getUniqueId().toString()) || playerBanned(player, claimTo)) && !hasAttacked && !hasTrust(player.getUniqueId(), claimTo)) {
-					if(!claimFrom.isWilderness()) {
-						if(playerBanned(player, claimFrom)) {
+				if (player.hasPermission("bfc.bypass") || player.getGameMode().equals(GameMode.SPECTATOR)) {
+					return;
+				}
+
+				if ((claimData.isAllBanned(claimTo.getUniqueId().toString()) || playerBanned(player, claimTo)) && !hasAttacked && !hasTrust(player.getUniqueId(), claimTo)) {
+					if (!claimFrom.isWilderness()) {
+						if (playerBanned(player, claimFrom)) {
 							final int sizeRadius = Math.max(claimTo.getLength(), claimTo.getWidth());
 
 							final Location greaterCorner = new Location(
@@ -72,47 +80,47 @@ public class GDListener implements Listener {
 
 							final LocationFinder lf = new LocationFinder(greaterCorner, lesserCorner, player.getWorld().getUID(), sizeRadius);
 							Bukkit.getScheduler().runTaskAsynchronously(BfcPlugin.getInstance(), () -> lf.IterateCircumferences(randomCircumferenceRadiusLoc -> {
-								if(randomCircumferenceRadiusLoc == null) {
-									if(Config.SAFE_LOCATION == null) { player.teleport(player.getWorld().getSpawnLocation()); }
-									else { player.teleport(Config.SAFE_LOCATION); }
+								if (randomCircumferenceRadiusLoc == null) {
+									if (Config.SAFE_LOCATION == null) {
+										player.teleport(player.getWorld().getSpawnLocation());
+									} else {
+										player.teleport(Config.SAFE_LOCATION);
+									}
+								} else {
+									player.teleport(randomCircumferenceRadiusLoc);
 								}
-								else { player.teleport(randomCircumferenceRadiusLoc);	}
 
 							}));
 
 						} else {
 							final Location tpLoc = player.getLocation().add(e.getFrom().toVector().subtract(e.getTo().toVector()).normalize().multiply(3));
 
-							if(tpLoc.getBlock().getType().equals(Material.AIR)) {
+							if (tpLoc.getBlock().getType().equals(Material.AIR)) {
 								player.teleport(tpLoc);
-							}
-							else {
+							} else {
 								final Location safeLoc = tpLoc.getWorld().getHighestBlockAt(tpLoc).getLocation().add(0D, 1D, 0D);
 								player.teleport(safeLoc);
 							}
 
-							if(e.getTo().getBlockX() == e.getFrom().getBlockX()) { ph.drawCircle(1, true); }
-							else { ph.drawCircle(1, false); }
+							ph.drawCircle(1, e.getTo().getBlockX() == e.getFrom().getBlockX());
 						}
 					} else {
 						final Location tpLoc = player.getLocation().add(e.getFrom().toVector().subtract(e.getTo().toVector()).normalize().multiply(3));
-						if(tpLoc.getBlock().getType().equals(Material.AIR)) { player.teleport(tpLoc); }
-						else {
+						if (tpLoc.getBlock().getType().equals(Material.AIR)) {
+							player.teleport(tpLoc);
+						} else {
 							final Location safeLoc = tpLoc.getWorld().getHighestBlockAt(tpLoc).getLocation().add(0D, 1D, 0D);
 							player.teleport(safeLoc);
 						}
 
-						if(e.getTo().getBlockX() == e.getFrom().getBlockX()) { ph.drawCircle(1, true); }
-						else { ph.drawCircle(1, false); }
+						ph.drawCircle(1, e.getTo().getBlockX() == e.getFrom().getBlockX());
 					}
 
-					if(!MessageHandler.spamMessageClaim.contains(player.getUniqueId().toString())) {
+					if (!MessageHandler.spamMessageClaim.contains(player.getUniqueId().toString())) {
 						MessageHandler.sendTitle(player, Messages.TITLE_MESSAGE, Messages.SUBTITLE_MESSAGE);
 						MessageHandler.spamMessageClaim.add(player.getUniqueId().toString());
 
-						Bukkit.getScheduler().runTaskLater(BfcPlugin.getInstance(), () -> {
-							MessageHandler.spamMessageClaim.remove(player.getUniqueId().toString());
-						}, 5L * 20L);
+						Bukkit.getScheduler().runTaskLater(BfcPlugin.getInstance(), () -> MessageHandler.spamMessageClaim.remove(player.getUniqueId().toString()), 5L * 20L);
 					}
 				}
 			}
@@ -122,10 +130,10 @@ public class GDListener implements Listener {
 	private boolean playerBanned(Player player, Claim claim) {
 		final ClaimData claimData = new ClaimData();
 		final String claimID = claim.getUniqueId().toString();
-		if(claimData.checkClaim(claimID)) {
-			if(claimData.bannedPlayers(claimID) != null) {
-				for(final String bp : claimData.bannedPlayers(claimID)) {
-					if(bp.equals(player.getUniqueId().toString())) {
+		if (claimData.checkClaim(claimID)) {
+			if (claimData.bannedPlayers(claimID) != null) {
+				for (final String bp : claimData.bannedPlayers(claimID)) {
+					if (bp.equals(player.getUniqueId().toString())) {
 						return true;
 					}
 				}
@@ -136,10 +144,11 @@ public class GDListener implements Listener {
 	}
 
 	private boolean hasTrust(UUID player, Claim claim) {
-		if(player.equals(claim.getOwnerUniqueId())) { return true; }
-		else if(claim.isUserTrusted(player, TrustTypes.MANAGER)) { return true; }
-		else if(claim.isUserTrusted(player, TrustTypes.BUILDER)) { return true; }
-		else { return false; }
+		if (player.equals(claim.getOwnerUniqueId())) {
+			return true;
+		} else if (claim.isUserTrusted(player, TrustTypes.MANAGER)) {
+			return true;
+		} else return claim.isUserTrusted(player, TrustTypes.BUILDER);
 	}
 
 }
