@@ -1,13 +1,12 @@
 package no.vestlandetmc.BanFromClaim.listener;
 
+import no.vestlandetmc.BanFromClaim.BfcPlugin;
+import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
+import no.vestlandetmc.BanFromClaim.utils.UpdateNotification;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
-
-import no.vestlandetmc.BanFromClaim.BfcPlugin;
-import no.vestlandetmc.BanFromClaim.handler.MessageHandler;
-import no.vestlandetmc.BanFromClaim.handler.UpdateNotification;
 
 public class PlayerListener implements Listener {
 
@@ -15,8 +14,8 @@ public class PlayerListener implements Listener {
 	public void playerJoin(PlayerJoinEvent p) {
 		final Player player = p.getPlayer();
 
-		if(player.isOp()) {
-			if(UpdateNotification.isUpdateAvailable()) {
+		if (player.isOp()) {
+			if (UpdateNotification.isUpdateAvailable()) {
 				MessageHandler.sendMessage(player, "&2" + BfcPlugin.getInstance().getDescription().getName() + " &ais outdated. Update is available!");
 				MessageHandler.sendMessage(player, "&aYour version is &2" + UpdateNotification.getCurrentVersion() + " &aand can be updated to version &2" + UpdateNotification.getLatestVersion());
 				MessageHandler.sendMessage(player, "&aGet the new update at &2https://www.spigotmc.org/resources/" + UpdateNotification.getProjectId());
