@@ -1,14 +1,13 @@
-package no.vestlandetmc.BanFromClaim.handler;
+package no.vestlandetmc.BanFromClaim.utils;
+
+import no.vestlandetmc.BanFromClaim.BfcPlugin;
+import org.bukkit.scheduler.BukkitRunnable;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.URL;
 import java.net.URLConnection;
-
-import org.bukkit.scheduler.BukkitRunnable;
-
-import no.vestlandetmc.BanFromClaim.BfcPlugin;
 
 public abstract class UpdateNotification extends BukkitRunnable {
 
@@ -26,15 +25,15 @@ public abstract class UpdateNotification extends BukkitRunnable {
 			final URL url = new URL("https://api.spigotmc.org/legacy/update.php?resource=" + projectId);
 			final URLConnection con = url.openConnection();
 
-			try(BufferedReader r = new BufferedReader(new InputStreamReader(con.getInputStream()))) {
+			try (BufferedReader r = new BufferedReader(new InputStreamReader(con.getInputStream()))) {
 				latestVersion = r.readLine();
 			}
 
-			if(isUpdateAvailable()) {
+			if (isUpdateAvailable()) {
 				onUpdateAvailable();
 			}
 
-		} catch(final IOException ex) {
+		} catch (final IOException ex) {
 			ex.getStackTrace();
 		}
 	}
